pyproject.toml
[project]
name = "powerpoint_automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "loguru>=0.7.3",
    "python-pptx>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
]